[ICU-900] Fixed some compiler warnings X-Commit-URL: https://ssl.icu-project.org/trac/changeset/4350

+++ icu4c/source/tools/gentest/gentest.c
@@ -113,7 +113,8 @@ createData(const char* outputDirectory) {
 
 
     if(dataLength!=(long)size) {
-        fprintf(stderr, "gentest: data length %ld != calculated size %lu\n", dataLength, size);
+        fprintf(stderr, "gentest: data length %ld != calculated size %lu\n",
+            dataLength, (unsigned long)size);
         exit(U_INTERNAL_PROGRAM_ERROR);
     }
 }
